+++ src/io.h
@@ -1,3 +1,3 @@
 #include "argparse.h"
 
-void read_file(struct options_t* args, double** input_vals);
+void read_file(struct options_t* args, float** input_vals);
