+++ src/main.cpp
@@ -18,14 +18,14 @@ int main(int argc, char **argv)
     get_opts(argc, argv, &opts);
 
     // Read the input file
-    double * vals;
+    float * vals;
     read_file(&opts, &vals);
 
     // Random assign Centroids
     // Set the seed for random.
     kmeans_set_rand_seed(opts.seed);
 
-    double * centroids = (double *) malloc(opts.num_cluster * opts.dims * sizeof(double));
+    float * centroids = (float *) malloc(opts.num_cluster * opts.dims * sizeof(float));
     for (int i = 0, index = 0; i < opts.num_cluster; i++){
         index = kmeans_rand() % opts.number_of_values;
         for (int j =0 ; j < opts.dims; j++){
@@ -52,7 +52,7 @@ int main(int argc, char **argv)
         for (int clusterId = 0; clusterId < opts.num_cluster; clusterId ++){
             printf("%d ", clusterId);
             for (int d = 0; d < opts.dims; d++){
-                printf("%0.5lf ", opts.centroids[clusterId*opts.dims  + d ]);
+                printf("%lf ", opts.centroids[clusterId*opts.dims  + d ]);
             }
             printf("\n");
         }
